Adds JSONL to SQLite parser (#4)
* Adds JSONL to SQLite parser

Not everything is properly unit-tested (particularly the utilities) and there's
much that I'd like to improve (e.g. to avoid the dependency on sqlitebiter,
which seemed like a good idea at first but has turned out to not be so great),
but deadlines are looming on the horizon, so I'll have to revisit later.

* Bug Fix: conversion was chocking when no output was specified.

@witchodealba reported that the following command was chocking:

./jsonl_to_sqlite.py resources/clients/20180704T161849_client_00002.jsonl.gz

The culprit was an additional extension being added to the input filename
that sqlitebiter received. Apparently sqlitebiter uses file extensions as
the only source to determine the parsing of input files, so it's easy to
get it confused if you give it a perfectly fine file with a nonmatching
extension.

While testing for this fix, I found that the error for nonexistent files
is not the most friendly, so I added a quick fix to make it more obvious,
but I think this will require a more thorough design review later.

* resolver dependencies version, sqlitebiter to 0.17.1

* Bug Fix: `file_util.add_extension_if_needed` fails in certain cases

The culprit was just a silly mistake. I added unit tests to avoid regressions.