Adds batching in `jsonl.write_to_file` (#5)
This should improve performance while writing to files, but it also allows us
to reuse `jsonl` code in `random_clients_generator`

The only difference from the previous implementation is that this one doesn't
include a newline at the end of the file. I checked the specification, and it
doesn't mandate that such a newline is present.