Client purchase order generation (#6)
* Adds JSONL to SQLite parser

Not everything is properly unit-tested (particularly the utilities) and there's much that I'd like to improve (e.g. to avoid the dependency on sqlitebiter, which seemed like a good idea at first but has turned out to not be so great), but deadlines are looming on the horizon, so I'll have to revisit later.

* Bug Fix: conversion was chocking when no output was specified.

@witchodealba reported that the following command was chocking:

./jsonl_to_sqlite.py resources/clients/20180704T161849_client_00002.jsonl.gz

The culprit was an additional extension being added to the input filename that sqlitebiter received. Apparently sqlitebiter uses file extensions as the only source to determine the parsing of input files, so it's easy to get it confused if you give it a perfectly fine file with a nonmatching extension.

While testing for this fix, I found that the error for nonexistent files is not the most friendly, so I added a quick fix to make it more obvious,
but I think this will require a more thorough design review later.

* Adds sell orders generator

This is a first stab at the sell orders generator. The core logic is as follows:

1. For each user, we determine (following a configurable probability distribution) whether it's an eager, casual or rare buyer.

2. We generate a series of purchases beginning at the user's registration date, whose spacing in time varies according to the selected user profile in the previous step

3. Each product is randomly picked from a sample of 100K diverse products

4. Price is sampled using a truncated log-normal distribution

5. Amount purchased is currently sampled from a uniform distribution, but it should probably come from a power law distribution.

* Fix broken `jsonl.write_to_file` function

The previous commit to resolve the merge conflicts to `master` failed to update a call to `open_file` (which used to be `_open_file`), which broke the `json.write_to_file` function.